Raw Weather without station_id to DB write dateUtil test

+++ src/main/scala/com/itechart/ny_accidents/entity/WeatherEntity.scala
@@ -6,9 +6,9 @@ import com.vividsolutions.jts.geom.Point
 case class WeatherEntity(id: Long,
                          stationId: Int,
                          localDateTime: Long,
-                         temperature: Int,
+                         temperature: Double,
                          pressure: Double,
-                         humidity: Int,
+                         humidity: Double,
                          windSpeed: Double,
                          phenomenon: String,
                          visibility: Double)
